split runner into subpackage
