+++ src/noob/__init__.py
@@ -1,9 +1,11 @@
+# ruff: noqa: I001 - import order meaningful to avoid cycles
+
 from noob.config import config as cfg
 from noob.cube import Cube
 from noob.logging import init_logger
 from noob.node import process_method
-from noob.runner import SynchronousRunner
 from noob.tube import Tube, TubeClassicEdition
+from noob.runner import SynchronousRunner
 from noob.types import Name
 
 __all__ = [
